requirements.txt
fastapi
uvicorn 
email-validator 
python-jose 
httpx 
sqlalchemy 
psycopg2 
passlib
python-dotenv
pytest
python-multipart
boto3